refactor(storage): move retry mocks
I will use them for more than just uploads.
